V1/requirements.txt
streamlit 
clip
torch
plotly
numpy 
pillow